add authentication

+++ database/migrations/2022_12_19_112402_create_katalog__obats_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->integer('stok');
             $table->integer('harga');
             $table->string('image');
-            $table->integer('kegunaan_id');
+            $table->unsignedBigInteger('kegunaan_id');
             $table->foreign('kegunaan_id')->references('id')->on('kegunaans');
             $table->timestamps();
         });
